pin wolfi glibc and related packages to workaround broke APKINDEXes (#10645)
For some reason Wolfi has been occasionally publishing APKINDEX files
where the latest version of ld-linux specifies a hard dependency on an
exact version of glibc which is not present in the APKINDEX yet. It goes
away within an hour usually, but it's frustrating in the meantime to not
be able to build the engine obviously.

For now, attempting to resolve this by just specifying glibc, ld-linux
and libcrypt1 versions that are one minor version ago. I repro'd the
problem by manually constructing an APKINDEX with the missing glibc
version (got the exact same error we see) and verified this approach
fixed it. We will see if it works in the real world too.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>